KQL search API addition, split of test cases
