+++ modules/version.json
@@ -1,3 +1,3 @@
 {
-    "FunctionVersion": "2.0.19"
+    "FunctionVersion": "2.0.20"
 }
